fuse: move to top level
