+++ fuse/dir.go
