+++ fuse/fuse.go
