+++ fuse/link.go
